rename: User -> Member로 이름변경
user가 mysql 예약어라서 중복돼서 바꿈
