+++ src/main/java/map/homepage/domain/member/MemberRepository.java
@@ -1,0 +1,11 @@
+package map.homepage.domain.member;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.Optional;
+
+public interface MemberRepository extends JpaRepository<Member,Long> {
+    Optional<Member> findByEmailAndProvider(String email, String provider);
+
+    Optional<Member> findByOauthId(String OauthId);
+}
